% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionTables.r
\name{selectionTable.basis}
\alias{selectionTable.basis}
\title{Makes calculations required by other functions ('basics'). Not exported.}
\usage{
selectionTable.basis(x, xd, name, vname, levls, levls.alt = levls,
  silent = FALSE)
}
\arguments{
\item{x}{sienaFit or sienaMeta: Results from SAOM}

\item{xd}{siena: the RSiena  Data Object or Data Set that
was used to generate x (if class(x) is sienaMeta, this must be one
of the Data Sets used to generate it, hopefully representative
w/r/t vname, and it cannot be a multigroup data set)}

\item{name}{character: the name of the endogenous variable
of interest (a network or behavior variable)}

\item{vname}{character: the actor variable name (i.e. predictor
variable)}

\item{levls}{numeric: number of ego levels in the variable given by vname}

\item{levls.alt}{numeric: number of alter levels in vname (usually = levls, in
which case you do not need to include it in the call)}
}
\value{
list: includes various calculations (predicted values for levels of
  vname, that kind of thing) needed for constructing the actual selection
  table.
}
\description{
Makes calculations required by other functions ('basics'). Not exported.
}
