% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionTables.r
\name{selectionTable.plot}
\alias{selectionTable.plot}
\title{Plots the selection table associated with a sienaFit object & other spex}
\usage{
selectionTable.plot(x, xd, name, vname, levls, levls.alt = levls,
  quad = TRUE, separation = 0, bw = FALSE, withMax = FALSE)
}
\arguments{
\item{x}{sienaFit: Results from a single group analysis}

\item{xd}{siena: the RSiena  Data Object or Data Set that
was used to generate x}

\item{name}{character: the name of the endogenous variable
of interest (a network or behavior variable)}

\item{vname}{character: the actor variable name (i.e. predictor
variable)}

\item{levls}{numeric: number of ego levels in the variable given by vname}

\item{levls.alt}{numeric (=levls): number of alter levels in vname
(usually = levls, in which case you do not need to include it in the call)}

\item{quad}{logical(=TRUE): If plotting a quadratic function, set to TRUE;
For similarity effects, set to FALSE.}

\item{separation}{numeric(=0): Higher values give the curves more separation
to improve readability. An advisable nonzero value is -0.01.}

\item{bw}{logical(=FALSE): If true, plot is black & white, otherwise color.}

\item{withMax}{logical (=FALSE): If TRUE, also plot the maximum over selection
probability over values of alters as a dashed line.}
}
\value{
A ggplot2 object
}
\description{
Plots the selection table associated with a sienaFit object & other spex
}
\details{

}
