% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionTables.r
\name{selectionTable}
\alias{selectionTable}
\title{Creates a selection (or influence) table (matrix)}
\usage{
selectionTable(x, xd, name, vname, levls, levls.alt = levls)
}
\arguments{
\item{x}{sienaFit: Results from a single group analysis}

\item{xd}{siena: the RSiena  Data Object or Data Set that
was used to generate x}

\item{name}{character: the name of the endogenous variable
of interest (a network or behavior variable)}

\item{vname}{character: the actor variable name (i.e. predictor
variable)}

\item{levls}{numeric: number of ego levels in the variable given by vname}

\item{levls.alt}{numeric: number of alter levels in vname (usually = levls, in
which case you do not need to include it in the call)}
}
\value{
matrix: the selection (or influence) table with alter values of
 vname across the columns, and ego values down the rows.
}
\description{
Creates a selection (or influence) table (matrix)
}
