% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionTables.r
\name{selectionTable.se}
\alias{selectionTable.se}
\title{Calculates standard errors for a linear combination
of elements of a selection table}
\usage{
selectionTable.se(x, xd, name, vname, levls, ww, levls.alt = levls)
}
\arguments{
\item{x}{sienaFit: Results from a single group analysis}

\item{xd}{siena: the RSiena  Data Object or Data Set that
was used to generate x}

\item{name}{character: the name of the endogenous variable
of interest (a network or behavior variable)}

\item{vname}{character: the actor variable name (i.e. predictor
variable)}

\item{levls}{numeric: number of ego levels in the variable given by vname}

\item{ww}{matrix: same format as the selection table of interest, containing
the coefficients of the linear combination}

\item{levls.alt}{numeric (=levls): number of alter levels in vname
(usually = levls, in which case you do not need to include it in the call)}
}
\value{
matrix: the selection (or influence) table se's, with alter values
 of 'vname' across the columns, and ego values down the rows.
}
\description{
Calculates standard errors for a linear combination
of elements of a selection table
}
\details{
The linear combination for which the standard error is computed is
sum_{h,k} ww[h,k] * selectionTable[h,k].
}
