% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionTables.r
\name{selectionTableWithMax}
\alias{selectionTableWithMax}
\title{Creates a dataframe with the selection table, plus max values
across alter values}
\usage{
selectionTableWithMax(x, xd, name, vname, levls, levls.alt = levls,
  discrete = TRUE)
}
\arguments{
\item{x}{sienaFit: Results from a single group analysis}

\item{xd}{siena: the RSiena  Data Object or Data Set that
was used to generate x}

\item{name}{character: the name of the endogenous variable
of interest (a network or behavior variable)}

\item{vname}{character: the actor variable name (i.e. predictor
variable)}

\item{levls}{numeric: number of ego levels in the variable given by vname}

\item{levls.alt}{numeric (=levls): number of alter levels in vname (usually = levls, in
which case you do not need to include it in the call)}
}
\value{
dataframe: MUST DOCUMENT THIS STRUCTURE WHEN YOU CAN FIGURE IT OUT!
 Apparently, there are two types of rows; if 'kind'=1, it's the selection
 table, and if 2, it's the max values.
}
\description{
Creates a dataframe with the selection table, plus max values
across alter values
}
\details{
Creates a data frame including the selection table for
siena data set xd, sienaFit object x,
actor covariate vname,
dependent variable name,
levels for ego levls, levels for alter levls.alt,
to which is appended the values for the maximum across alter,
where alter ranges over levls.alt (a grid with 200 points between
minimum and maximum is used).
}
